add flyway and clean tests

+++ src/main/java/dk/cngroup/kolnicka/Wish.java
@@ -22,5 +22,5 @@ public class Wish {
     private Long id;
     private String description;
     private String link;
-    private long priority;
+    private Integer priority;
 }
